Gadget : Defer signal creation until first access
Signals are large and slow to construct, and for many Gadgets are never actually accessed. Deferring creation reduces GraphGadget construction time for a large graph by 25%.

This also appears to give a 30% reduction in draw times for the same GraphGadget, presumably by giving us better memory locality for the things we actually need in `Gadget::render()`.