GafferUITest : Fix inconsistent UI state in BackgroundMethod test
BackgroundMethod test cases made use of a `WaitingSlot` to pause
the test thread until some background method had been called.

On some platforms, in some environments, the window holding the
TestWidget may, or may not be visible on screen by the time a wait was
requested. If the widget window is visible, then the pre-call of
`setEnabled( False )` causes a `valueChangedSignal` of reason
`Edit` to be emit.

The previous `WaitingSlot` did not discriminate what reason for
valueChanged was being awaited. This resulted in the wait exiting
before the background method had run when the window was on screen,
causing premature continuation of the test case, and consequent
failure of some assert later on.

This caused platform and/or number-of-tests dependent failures.

This commit ensures the window is visible, and changes to a more
specific wait behaviour. We also remove public visibility of the internal
test classes, they were never intended to be used outside this module.