Gaffer.bat : disable Jemalloc.bat by default (Windows only)
It seems Jemalloc is not tested extensively on Windows (according to Github discussions including the developers) and Github issues suggest stability across major versions can be spotty. DCCs like Houdini also only use Jemalloc on Linux so we're in good company.

The 5.2.1 (Gaffer is on 3.6 for stability reasons) release focuses on Windows so perhaps in future upgrades this can be enabled by default.